[FIX] SettingsPlugin: move dispatch logic in CellPlugin
Task: 4908195
